Moved Hex To Enviornment (In Plugins)
So every plugin set needs its own enviornment to live in, and that will be developed over in the plugin area, in an "Enviornment" folder.  Also, adjusted the compiler to ignore the new folder in "Plugins"
