+++ Main/dataStructure/Hex.h
@@ -160,15 +160,6 @@ private:
      std::map< const std::size_t, std::map<Entity* const, Enticap*>>* >> nearBy;
 };
 
-class Hex : public HexInternals
-{
-public:
-    //Give it the matrix coordinates, and it will generate pixel coordinates
-    Hex(const unsigned int col, const unsigned int row, const double hexRadius);
-
-    virtual void update(unsigned int resolution) override;
-//private:
-
-};
+#include "../../Plugins/Enviornment/Hex.h"
 
 #endif
