+++ Plugins/Enviornment/Hex.cpp
@@ -1,0 +1,5 @@
+//Give it the matrix coordinates, and it will generate pixel coordinates
+Bin::Hex::Hex(const unsigned int col, const unsigned int row, const double hexRadius)
+ : HexInternals(col, row, hexRadius) {}
+
+void Bin::Hex::update(unsigned int resolution) {if (resolution) {}}
